requirements.txt
tensorflow
pyttsx3
imutils
cv2
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy
Pillow
scipy>=0.17.0
